Trying out userSchema that has LL property
