+++ models/user.js
@@ -7,7 +7,10 @@ const userSchema = new mongoose.Schema({
   lastName: String,
   username: {type: String, required: true},
   password: {type: String, required: true},
-  questions: [{imageURL: String, question: String, answer: String, M: Number}]
+  // questions: [{imageURL: String, question: String, answer: String, M: Number}]
+  questions: {
+    head: { type: Object }
+  }
 });
 
 userSchema.set('toObject', {
