+++ routes/users.js
@@ -2,8 +2,18 @@
 
 const express = require('express');
 const User = require('../models/user');
+const LinkedList = require('../utils/LinkedList');
 const Router = express.Router();
 const seedQuestions = require('../db/seed/questions');
+
+function createLinkedList() {
+  // this only creates the new LinkedList from scratch, does not grab the local question data per user
+  const questionLinkedList = new LinkedList();
+  for (let i = 0; i < seedQuestions.length; i++) {
+    questionLinkedList.insertFirst(seedQuestions[i]);
+  }
+  return questionLinkedList;
+}
 
 Router.post('/', (req, res, next) => {
   const { firstName, lastName, username, password } = req.body;
@@ -74,7 +84,8 @@ Router.post('/', (req, res, next) => {
   if (lastName) {
     last = lastName.trim();
   }
-  console.log(seedQuestions);
+  const questions = createLinkedList();
+  console.log(questions);
 
   return User.hashPassword(password)
     .then(digest => {
@@ -83,7 +94,7 @@ Router.post('/', (req, res, next) => {
         lastName: last,
         username,
         password: digest,
-        questions: seedQuestions
+        questions
       };
       return User.create(newUser);
     })
